# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Real-Time Analytics Pipeline API",
    "version": "1.0.0",
    "description": "Enterprise-grade real-time analytics pipeline API. Supports event ingestion, analytics queries, dashboard management, and real-time WebSocket streaming.",
    "contact": {
      "name": "Engineering Team",
      "email": "engineering@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://api.example.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "VALIDATION_ERROR" },
              "message": { "type": "string", "example": "Validation failed" },
              "details": { "type": "object", "example": {} }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
...
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "minLength": 3, "maxLength": 100, "example": "admin" },
          "password": { "type": "string", "minLength": 8, "maxLength": 128, "example": "securepass123!" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "email", "password"],
        "properties": {
          "username": { "type": "string", "minLength": 3, "maxLength": 100, "pattern": "^[a-zA-Z0-9_]+$", "example": "admin" },
          "email": { "type": "string", "format": "email", "example": "admin@example.com" },
          "password": { "type": "string", "minLength": 8, "maxLength": 128, "example": "securepass123!" }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "refresh_token": { "type": "string" },
          "token_type": { "type": "string", "example": "bearer" },
          "expires_in": { "type": "integer", "example": 900 }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "user_id": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin", "analyst", "viewer"] },
          "is_active": { "type": "boolean" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register_user",
        "summary": "Register a new user",
        "description": "Creates a new user account and returns JWT tokens.",
        "tags": ["Authentication"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "422": {
            "description": "Validation error (username or email taken)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_user",
        "summary": "Login",
        "description": "Authenticates a user and returns JWT tokens.",
        "tags": ["Authentication"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_token",
        "summary": "Refresh access token",
        "description": "Exchanges a refresh token for a new access token and refresh token.",
        "tags": ["Authentication"],
        "security": [],
        "parameters": [
          {
            "name": "X-Refresh-Token",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "get_current_user",
        "summary": "Get current user",
        "description": "Returns the currently authenticated user's profile.",
        "tags": ["Authentication"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user information",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}